+++ resourcet/Data/Acquire.hs
@@ -5,7 +5,6 @@
 module Data.Acquire
     ( Acquire
     , with
-    , withEx
     , mkAcquire
     , mkAcquireType
     , allocateAcquire
@@ -13,15 +12,9 @@ module Data.Acquire
     ) where
 
 import Control.Monad.Trans.Resource.Internal
-import Control.Monad.Trans.Resource
 import Data.Acquire.Internal
-import Control.Applicative (Applicative (..))
-import Control.Monad.Base (MonadBase (..))
 import Control.Monad.IO.Class (MonadIO (..))
-import Control.Monad.Trans.Control (MonadBaseControl, control)
-import qualified Control.Exception.Lifted as E
-import Data.Typeable (Typeable)
-import Control.Monad (liftM, ap)
+import qualified Control.Exception as E
 
 -- | Allocate a resource and register an action with the @MonadResource@ to
 -- free the resource.