+++ resourcet/Data/Acquire/Internal.hs
@@ -9,21 +9,17 @@ module Data.Acquire.Internal
     ( Acquire (..)
     , Allocated (..)
     , with
-    , withEx
     , mkAcquire
     , ReleaseType (..)
     , mkAcquireType
     ) where
 
 import Control.Applicative (Applicative (..))
-import Control.Monad.Base (MonadBase (..))
-import Control.Monad.IO.Class (MonadIO (..))
-import Control.Monad.Trans.Control (MonadBaseControl, control)
-import qualified Control.Exception.Lifted as E
+import Control.Monad.IO.Unlift (MonadIO (..), MonadUnliftIO, withRunInIO)
+import qualified Control.Exception as E
 import Data.Typeable (Typeable)
 import Control.Monad (liftM, ap)
-import qualified Control.Monad.Catch as C
-import GHC.IO (unsafeUnmask)
+import qualified Control.Monad.Catch as C ()
 
 -- | The way in which a release is called.
 --
@@ -69,9 +65,6 @@ instance MonadIO Acquire where
         x <- restore f
         return $! Allocated x (const $ return ())
 
-instance MonadBase IO Acquire where
-    liftBase = liftIO
-
 -- | Create an @Acquire@ value using the given allocate and free functions.
 --
 -- Since 1.1.0
@@ -101,39 +94,12 @@ mkAcquireType create free = Acquire $ \restore -> do
 -- @bracket@.
 --
 -- Since 1.1.0
-with :: MonadBaseControl IO m
+with :: MonadUnliftIO m
      => Acquire a
      -> (a -> m b)
      -> m b
-with (Acquire f) g = control $ \run -> E.mask $ \restore -> do
+with (Acquire f) g = withRunInIO $ \run -> E.mask $ \restore -> do
     Allocated x free <- f restore
     res <- restore (run (g x)) `E.onException` free ReleaseException
     free ReleaseNormal
     return res
-
--- | Same as @with@, but uses the @MonadMask@ typeclass from exceptions instead
--- of @MonadBaseControl@ from exceptions.
---
--- Since 1.1.3
-#if MIN_VERSION_exceptions(0,6,0)
-withEx :: (C.MonadMask m, MonadIO m)
-#else
-withEx :: (C.MonadCatch m, MonadIO m)
-#endif
-       => Acquire a
-       -> (a -> m b)
-       -> m b
-withEx (Acquire f) g = do
-    -- We need to do some funny business, since the restore we get below is
-    -- specialized to the m from the result, whereas we need a restore function
-    -- in IO. Checking the current masking state is exactly how mask is
-    -- implemented in base.
-    origMS <- liftIO E.getMaskingState
-
-    C.mask $ \restore -> do
-        Allocated x free <- liftIO $ f $ case origMS of
-            E.Unmasked -> unsafeUnmask
-            _ -> id
-        res <- restore (g x) `C.onException` liftIO (free ReleaseException)
-        liftIO $ free ReleaseNormal
-        return res